+++ pkg/querier/blocks_store_queryable.go
@@ -24,6 +24,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
+	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/thanos-io/thanos/pkg/block"
@@ -98,6 +99,7 @@ type BlocksStoreClient interface {
 type BlocksStoreLimits interface {
 	bucket.TenantConfigProvider
 
+	MaxLabelsQueryLength(userID string) time.Duration
 	MaxChunksPerQueryFromStore(userID string) int
 	StoreGatewayTenantShardSize(userID string) int
 }
@@ -351,6 +353,16 @@ func (q *blocksStoreQuerier) LabelNames(matchers ...*labels.Matcher) ([]string, 
 
 	minT, maxT := q.minT, q.maxT
 
+	level.Debug(spanLog).Log("start", util.TimeFromMillis(minT).UTC().String(), "end",
+		util.TimeFromMillis(maxT).UTC().String(), "matchers", util.MatchersStringer(matchers))
+
+	{
+		// Clamp max time range.
+		startTime, endTime := model.Time(minT), model.Time(maxT)
+		maxQueryLength := q.limits.MaxLabelsQueryLength(q.userID)
+		minT = int64(clampTime(spanCtx, startTime, maxQueryLength, endTime.Add(-maxQueryLength), true, "start", "max label query length", spanLog))
+	}
+
 	var (
 		resMtx            sync.Mutex
 		resNameSets       = [][]string{}
@@ -385,6 +397,16 @@ func (q *blocksStoreQuerier) LabelValues(name string, matchers ...*labels.Matche
 	defer spanLog.Span.Finish()
 
 	minT, maxT := q.minT, q.maxT
+
+	level.Debug(spanLog).Log("start", util.TimeFromMillis(minT).UTC().String(), "end",
+		util.TimeFromMillis(maxT).UTC().String(), "matchers", util.MatchersStringer(matchers))
+
+	{
+		// Clamp max time range.
+		startTime, endTime := model.Time(minT), model.Time(maxT)
+		maxQueryLength := q.limits.MaxLabelsQueryLength(q.userID)
+		minT = int64(clampTime(spanCtx, startTime, maxQueryLength, endTime.Add(-maxQueryLength), true, "start", "max label query length", spanLog))
+	}
 
 	var (
 		resValueSets = [][]string{}
@@ -423,10 +445,7 @@ func (q *blocksStoreQuerier) selectSorted(sp *storage.SelectHints, matchers ...*
 	spanLog, spanCtx := spanlogger.NewWithLogger(q.ctx, q.logger, "blocksStoreQuerier.selectSorted")
 	defer spanLog.Span.Finish()
 
-	minT, maxT := q.minT, q.maxT
-	if sp != nil {
-		minT, maxT = sp.Start, sp.End
-	}
+	minT, maxT := sp.Start, sp.End
 
 	var (
 		convertedMatchers = convertMatchersToLabelMatcher(matchers)